fix(search): Sanity 直叩きの CORS エラーを Server-side proxy で根本解消
問題: Vercel preview から Sanity API を直接叩くと
  Access-Control-Allow-Origin がなく CORS で全 fetch が拒否される

対応: /api/search Route Handler を新規作成し、
- Sanity 取得ロジックを useSearch.ts から移植（サーバー側で実行）
- 検索専用 client は useCdn:true で Sanity CDN 経由（高速化）
- Cache-Control: s-maxage=300, stale-while-revalidate=600 で
  Vercel Edge にもキャッシュ
- クライアント（useSearch.ts）は fetch('/api/search') に統一
  → 同一オリジン、CORS 不要
- React Query staleTime 5分 + Vercel s-maxage 5分 で二重キャッシュ

副次効果: 検索ページの体感速度が大幅改善（CDN + キャッシュ層）
